aider: Implemented print statements for the "Jump to address" and "Halt" conditions in the handling of the terminating block.
# Aider chat conversation:

USER:
There is still one open TODO item:

```rust
        // Check for end-of-input condition
        if byte_count == 6 {
            if load_address % 2 == 0 {
                // Jump to address (not implemented)
            } else {
                // Halt (not implemented)
            }
            break;

ASSISTANT: To address the TODO item, we need to implement the functionality for the "Jump to address" and "Halt" conditions. Since the specifics of how to handle these conditions are not provided, I will add a simple print statement for each to indicate that the condition has been reached. If you have specific logic that needs to be implemented for these conditions, please provide the details.

Here's the change to implement print statements for the "Jump to address" and "Halt
^C KeyboardInterrupt

USER:         // Check for end-of-input condition
        if byte_count == 6 {
            if load_address % 2 == 0 {
                // Jump to address (not implemented)
            } else {
                // Halt (not implemented)
            }
            break;
        }

ASSISTANT: To address the TODO item, we will implement simple print statements for the "Jump to address" and "Halt" conditions.
^C KeyboardInterrupt

USER: There is still one open TODO item: The "Jump to address" and "Halt" branches of the handling of the terminating block. We should just print out whether the loader will jump that address or not.

ASSISTANT: Understood. We will add print statements to indicate whether the loader would jump to the address or halt based on the `load_address`. Here's the change for the `ptap/src/main.rs` file:

```rust
ptap/src/main.rs
<<<<<<< SEARCH
        // Check for end-of-input condition
        if byte_count == 6 {
            if load_address % 2 == 0 {
                // Jump to address (not implemented)
            } else {
                // Halt (not implemented)
            }
            break;
        }
=======
        // Check for end-of-input condition
        if byte_count == 6 {
            if load_address % 2 == 0 {
                println!("Jump to address: 0x{:04X}", load_address);
            } else {
                println!("Halt at address: 0x{:04X}", load_address);
            }
            break;
        }
>>>>>>> REPLACE
```

This change will print a message indicating the action to be taken ("Jump to address" or "Halt") when the end-of-input condition is detected with a byte count of 6. If the `load_address` is even, it will print the jump message; if it's odd, it will print the halt message.